Update A9MEM1570.md
This is Zigbee green power device. I've included info about it copied from another device that works in that standard (https://www.zigbee2mqtt.io/devices/8718696743133.html). Please include this as it was confusing for me that the device listed on your site wasn't working in my network. I later read about this substandard, and wanted everyone to know about it.